iris-django/handballStats/apiConnection/serializers.py
from rest_framework import serializers
from .models import Team

class TeamSerializer(serializers.ModelSerializer):
    class Meta:
        model = Team
        fields = ('name', 'wins', 'opponents', 'last_match', 'last_win', 'added')