iris-django/handballStats/requirements.txt
asgiref==3.3.1
Django==3.1.4
djangorestframework==3.12.2
gunicorn==20.0.4
psycopg2-binary==2.8.6
pytz==2020.4
sqlparse==0.4.1
